+++ services/yacynode/internal/documentsearch/term_posting.go
@@ -4,6 +4,6 @@ import "github.com/nikitakarpei/yacy-rwi-node/yacymodel"
 
 type termPosting struct {
 	documentIdentifier yacymodel.Hash
-	occurrences        uint64
-	textPosition       uint64
+	occurrences        int
+	textPosition       int
 }
